refactor: Extract parsing strategies from FastDateParser
- Extracted `CaseInsensitiveTextStrategy` for case-insensitive text parsing.
- Extracted `CopyQuotedDateParsingStrategy` for quoted/static field parsing.
- Extracted `DateParsingStrategy` as the base class for all strategies.
- Extracted `ISO8601TimeZoneDateParsingStrategy` for ISO 8601 time zone parsing.
- Extracted `NumberDateParsingStrategy` for numeric field parsing.
- Extracted `TimeZoneDateParsingStrategy` for general time zone parsing.

This change improves modularity, readability, and maintainability by adhering to the Single Responsibility Principle.
